+++ config/i18n.json
@@ -976,5 +976,12 @@
     "Num": "{{num, number}} rolls in your wish list",
     "UpdateExternalSource": "Update Wish List Source",
     "WishListNotes": "Wish List Notes: {{notes}}"
+  },
+  "ActiveMode": {
+    "ButtonOn": "Active Mode",
+    "ButtonOff": "Inventory Mode",
+    "Beta": "BETA",
+    "Pursuits": "Pursuits",
+    "NoQuests": "You're not tracking any quests"
   }
 }
