+++ src/app/active-mode/Views/current-activity/util.ts
